Bugfix: Added nullptr check to Process
In case a process was already running QProcess would reference
to a nullptr. We need to perform a check if it started before
sending a kill signal.

+++ src/syncconnector.cpp
@@ -276,7 +276,10 @@ void SyncConnector::onSslError(QNetworkReply* reply)
   
 SyncConnector::~SyncConnector()
 {
-  mpSyncProcess->kill();
+  if (mpSyncProcess != nullptr)
+  {
+    mpSyncProcess->kill();
+  }
 }
   
 } // connector
